# Improved performance of concatenating non-aligned validities (15x)