@ritchie46 2021-08-17
Love it! I see some interesting bit comments. Have you got a summary of what you do? A `memcpy` instead of iterators?