@sundy-li 2021-08-24
A bug found in https://github.com/jorgecarleitao/arrow2/pull/325